fix recursive attribute

+++ modules/File.js
@@ -30,13 +30,15 @@ module.exports = class File {
         }
     }
 
-    async rmDir(path = false, config = { recursive: false }) {
+    async rmDir(path = false, config = { recursive: true }) {
         if(!path) throw new Error('Incorrected path');
 
         await rm(path);
 
         async function rm(path) {
             fs.readdirSync(path).forEach(async file => {
+                if(config.recursive === false) throw new Error(`${ path } is not empty`); 
+                
                 if(Path.extname(file)) {
                     fs.unlinkSync(`${ path }/${ file }`, err => console.log(err));
                 } else {
